Create clinical summary package & separate classes
