+++ src/main/scala/com/github/vitalsoftware/scalaredox/models/RedoxMessage.scala
@@ -44,7 +44,7 @@ sealed trait VisitRedoxMessage extends PatientRedoxMessage with HasVisit
  */
 @jsonDefaults case class Visit(
   Meta: Meta,
-  Header: ClinicalSummaryHeader,
+  Header: clinicalsummary.Header,
   AllergyText: Option[String] = None,
   Allergies: Seq[Allergy] = Seq.empty,
   AssessmentText: Option[String] = None,
