+++ src/main/scala/com/github/vitalsoftware/scalaredox/models/clinicalsummary/ClinicalSummaryTraits.scala
@@ -1,0 +1,11 @@
+package com.github.vitalsoftware.scalaredox.models.clinicalsummary
+
+import com.github.vitalsoftware.scalaredox.models._
+
+trait ClinicalSummaryLike {
+  def Meta: Meta
+}
+
+trait HasClinicalSummaryPatient {
+  def Patient: Patient
+}
